[Update] UniWinAPi .cs で自ウィンドウ取得を Process.GetCurrentProcess() に基づくものに変更。 [Change] コマンドライン引数でVRMファイル指定を受け付けていたが、オプションを除外できていないことで問題があったため、引数は扱わないものとした。

+++ Assets/UniWinApi/Scripts/Wrappers/WinApi.cs
@@ -108,7 +108,7 @@ public class WinApi {
 	public static extern int GetClassName(IntPtr hWnd, StringBuilder lpClassName, int nMaxCount);
 	
 	[DllImport("user32.dll")]
-	public static extern uint GetWindowThreadProcessId (IntPtr hWnd, out long lpdwProcessId);
+	public static extern uint GetWindowThreadProcessId (IntPtr hWnd, out int lpdwProcessId);
 	
 	[DllImport("user32.dll")]
 	public static extern IntPtr FindWindow (string lpszClass, string lpszTitle);
